+++ src/GEval/EvaluationScheme.hs
@@ -1,22 +1,33 @@
 module GEval.EvaluationScheme
-  (EvaluationScheme(..), evaluationSchemeMetric, applyPreprocessingOperations, evaluationSchemeName, PreprocessingOperation(..))
+  (EvaluationScheme(..),
+   evaluationSchemeMetric,
+   applyPreprocessingOperations,
+   evaluationSchemeName,
+   evaluationSchemePriority,
+   PreprocessingOperation(..))
   where
 
 import GEval.Metric
 
 import Text.Regex.PCRE.Heavy
 import Text.Regex.PCRE.Light.Base (Regex(..))
-import Data.Text (Text(..), concat, toLower, toUpper, pack, unpack)
-import Data.List (intercalate, break)
+import Data.Text (Text(..), concat, toLower, toUpper, pack, unpack, words, unwords)
+import Data.List (intercalate, break, sort)
 import Data.Either
-import Data.Maybe (fromMaybe)
+import Data.Maybe (fromMaybe, catMaybes)
 import qualified Data.ByteString.UTF8 as BSU
 
 
 data EvaluationScheme = EvaluationScheme Metric [PreprocessingOperation]
   deriving (Eq)
 
-data PreprocessingOperation = RegexpMatch Regex | LowerCasing | UpperCasing | SetName Text
+data PreprocessingOperation = RegexpMatch Regex
+                              | LowerCasing
+                              | UpperCasing
+                              | Sorting
+                              | SetName Text
+                              | SetPriority Int
+                              | RegexpSubstition Regex Text
   deriving (Eq)
 
 leftParameterBracket :: Char
@@ -39,16 +50,37 @@ readOps ('l':theRest) = (LowerCasing:ops, theRest')
 readOps ('u':theRest) = (UpperCasing:ops, theRest')
     where (ops, theRest') = readOps theRest
 readOps ('m':theRest) = handleParametrizedOp (RegexpMatch . (fromRight undefined) . ((flip compileM) []) . BSU.fromString) theRest
+readOps ('S':theRest) = (Sorting:ops, theRest')
+    where (ops, theRest') = readOps theRest
 readOps ('N':theRest) = handleParametrizedOp (SetName . pack) theRest
+readOps ('P':theRest) = handleParametrizedOp (SetPriority . read) theRest
+readOps ('s':theRest) = handleParametrizedBinaryOp (\a b -> RegexpSubstition (fromRight undefined $ compileM (BSU.fromString a) []) (pack b)) theRest
 readOps s = ([], s)
 
 handleParametrizedOp :: (String -> PreprocessingOperation) -> String -> ([PreprocessingOperation], String)
-handleParametrizedOp constructor (leftParameterBracket:theRest) =
+handleParametrizedOp constructor theRest =
+  case parseParameter theRest of
+    (Nothing, s) -> ([], s)
+    (Just param, theRest') -> let (ops, theRest'') = readOps theRest'
+                             in ((constructor param):ops, theRest'')
+
+handleParametrizedBinaryOp :: (String -> String -> PreprocessingOperation) -> String -> ([PreprocessingOperation], String)
+handleParametrizedBinaryOp constructor theRest =
+  case parseParameter theRest of
+    (Nothing, s) -> ([], s)
+    (Just paramA, theRest') ->
+      case parseParameter theRest' of
+        (Nothing, s) -> ([], s)
+        (Just paramB, theRest'') -> let (ops, theRest''') = readOps theRest''
+                                   in ((constructor paramA paramB):ops, theRest''')
+
+parseParameter :: String -> (Maybe String, String)
+parseParameter (leftParameterBracket:theRest) =
   case break (== rightParameterBracket) theRest of
-    (s, []) -> ([], s)
-    (param, (_:theRest')) -> let (ops, theRest'') = readOps theRest'
-                            in ((constructor param):ops, theRest'')
-handleParametrizedOp _ s = ([], s)
+    (s, []) -> (Nothing, s)
+    (param, (_:theRest')) -> (Just param, theRest')
+parseParameter s = (Nothing, s)
+
 
 instance Show EvaluationScheme where
   show (EvaluationScheme metric operations) = (show metric) ++ (if null operations
@@ -58,10 +90,21 @@ instance Show EvaluationScheme where
 evaluationSchemeName :: EvaluationScheme -> String
 evaluationSchemeName scheme@(EvaluationScheme metric operations) = fromMaybe (show scheme) (findNameSet operations)
 
+evaluationSchemePriority scheme@(EvaluationScheme _ operations) = fromMaybe defaultPriority (findPrioritySet operations)
+  where defaultPriority = 1
+
 findNameSet :: [PreprocessingOperation] -> Maybe String
-findNameSet [] = Nothing
-findNameSet ((SetName name):_) = Just (unpack name)
-findNameSet (_:ops) = findNameSet ops
+findNameSet ops = case names of
+  [] -> Nothing
+  _ -> Just $ intercalate " " names
+  where names = catMaybes $ map extractName ops
+        extractName (SetName n) = Just (unpack n)
+        extractName _ = Nothing
+
+findPrioritySet :: [PreprocessingOperation] -> Maybe Int
+findPrioritySet [] = Nothing
+findPrioritySet ((SetPriority p):_) = Just p
+findPrioritySet (_:ops) = findPrioritySet ops
 
 evaluationSchemeMetric :: EvaluationScheme -> Metric
 evaluationSchemeMetric (EvaluationScheme metric _) = metric
@@ -70,10 +113,31 @@ instance Show PreprocessingOperation where
   show (RegexpMatch (Regex _ regexp)) = parametrizedOperation "m" (BSU.toString regexp)
   show LowerCasing = "l"
   show UpperCasing = "u"
+  show Sorting = "S"
   show (SetName t) = parametrizedOperation "N" (unpack t)
+  show (SetPriority p) = parametrizedOperation "P" (show p)
+  show (RegexpSubstition (Regex _ regexp) s) = "s" ++ (formatParameter $ BSU.toString regexp) ++ (formatParameter $ unpack s)
+
+applySubstitution :: Regex -> Text -> Text -> Text
+applySubstitution r substitution t =
+  gsub r (handleRefs substitution) t
+
+handleRefs :: Text -> Text -> [Text] -> Text
+handleRefs substitution mainMatch subMatches = gsub refRegexp handleRef substitution
+  where Right refRegexp = compileM (BSU.fromString "\\\\\\d+") []
+        indexables = mainMatch : subMatches
+        handleRef :: Text -> Text
+        handleRef ref =
+          let ix = (read $ tail $ unpack ref)
+          in if ix >= length indexables
+             then (pack "")
+             else indexables !! ix
 
 parametrizedOperation :: String -> String -> String
-parametrizedOperation opCode opArg = opCode ++ [leftParameterBracket] ++ opArg ++ [rightParameterBracket]
+parametrizedOperation opCode opArg = opCode ++ (formatParameter opArg)
+
+formatParameter :: String -> String
+formatParameter p = [leftParameterBracket] ++ p ++ [rightParameterBracket]
 
 applyPreprocessingOperations :: EvaluationScheme -> Text -> Text
 applyPreprocessingOperations (EvaluationScheme _ operations) t = foldl (flip applyPreprocessingOperation) t operations
@@ -82,4 +146,7 @@ applyPreprocessingOperation :: PreprocessingOperation -> Text -> Text
 applyPreprocessingOperation (RegexpMatch regex) = Data.Text.concat . (map fst) . (scan regex)
 applyPreprocessingOperation LowerCasing = toLower
 applyPreprocessingOperation UpperCasing = toUpper
+applyPreprocessingOperation Sorting = Data.Text.unwords . sort . Data.Text.words
 applyPreprocessingOperation (SetName _) = id
+applyPreprocessingOperation (SetPriority _) = id
+applyPreprocessingOperation (RegexpSubstition regex substition) = applySubstitution regex substition